+++ CHANGELOG.md
@@ -1,3 +1,5 @@
 # [1.0.0] **Ristretto"
 
 - Production-ready release of CupTrack - digital coffe fund
+- PIN-change for user on terminal
+- create user on terminal